#50 Refactor project structure to HEX-DDD
