+++ public/index.php
@@ -2,7 +2,7 @@
 require __DIR__ . '/../vendor/autoload.php';
 
 // Instantiate the app
-$app = \Skeleton\App\AppFactory::createApp();
+$app = \Skeleton\Application\AppFactory::createApp();
 
 // Run app
 $app->run();
