--- src/Entity/TaskList.php
@@ -1,33 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Skeleton\Entity;
-
-final class TaskList
-{
-    private ?int $id;
-    private string $title;
-
-    public function __construct(string $title)
-    {
-        $this->title = $title;
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id ?? null;
-    }
-
-    public function getTitle(): string
-    {
-        return $this->title;
-    }
-
-    public function setTitle(string $title): self
-    {
-        $this->title = $title;
-
-        return $this;
-    }
-}
